requirements.txt
aiogram>=3.2.0,<4.0.0
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.6.0,<3.0.0
aiosqlite>=0.20.0,<0.21.0
typing-extensions>=4.8.0
aiohttp>=3.9.0,<4.0.0
uvloop>=0.19.0; sys_platform == 'linux'